Refactor: Reorganize `Shiritori` library file structure. * Add folders for each class that export the class, interface, and additional types.
