+++ src/libs/shiritori/character/index.ts
@@ -1,0 +1,4 @@
+import { Character } from "./character";
+import { iCharacter } from "./interfaces/character";
+
+export { Character, iCharacter };
