+++ src/libs/shiritori/deck/index.ts
@@ -1,0 +1,5 @@
+import Deck from "./deck";
+import { iDeck } from "./interfaces/deck";
+
+export default Deck;
+export { iDeck };
