+++ src/libs/shiritori/index.ts
@@ -1,2 +1,5 @@
 import Shiritori from "./shiritori";
+import { iShiritori } from "./interfaces/shiritori";
+
 export default Shiritori;
+export { iShiritori };
