+++ src/libs/shiritori/player/index.ts
@@ -1,0 +1,7 @@
+import Player, { PlayerType } from "./player";
+import { iPlayer } from "./interfaces/player";
+import Computer from "./computer";
+import User from "./user";
+
+export default Player;
+export { Computer, User, iPlayer, PlayerType };
