+++ src/libs/shiritori/word/index.ts
@@ -1,0 +1,5 @@
+import Word from "./word";
+import { iWord } from "./interfaces/word";
+
+export default Word;
+export { iWord };
