+++ src/libs/shiritori/word/interfaces/word.ts
